requirements.txt
aiohttp
aiohttp_jinja2
jinja2
aiopg.sa
aiofiles
pathlib
yaml
